Subnetting
This is a basic design of a network with 3 subnets. 1 private subnet that can only be reached from within the LAN. This subnet can host up to 28 hosts. 1 private subnet that has access to the Internet through a NAT gateway. This subnet must can accomodate up to 60 hosts. 1 public subnet with an internet gateway. This subnet can accomodatre up to 12 hosts. There is NAT placed withing the public subnet and the public subnet is connected to the internet through the internet gateway. A firewall wall is installed to protect the LAN network.